requirements.txt
streamlit
pandas
python-dotenv
azure-ai-formrecognizer
azure-storage-blob
azure-core